implemented feeback of calv
